Move protocol parent states into appropriate module
