+++ swap/src/protocol/bob/swap.rs
@@ -1,7 +1,7 @@
 use anyhow::{bail, Result};
 use async_recursion::async_recursion;
 use rand::{CryptoRng, RngCore};
-use std::{fmt, sync::Arc};
+use std::sync::Arc;
 use tokio::select;
 use tracing::info;
 use uuid::Uuid;
@@ -12,44 +12,6 @@ use crate::{
     protocol::bob::{self, event_loop::EventLoopHandle, state::*},
     ExpiredTimelocks, SwapAmounts,
 };
-
-#[derive(Debug, Clone)]
-pub enum BobState {
-    Started {
-        state0: State0,
-        amounts: SwapAmounts,
-    },
-    Negotiated(State2),
-    BtcLocked(State3),
-    XmrLocked(State4),
-    EncSigSent(State4),
-    BtcRedeemed(State5),
-    CancelTimelockExpired(State4),
-    BtcCancelled(State4),
-    BtcRefunded(State4),
-    XmrRedeemed,
-    BtcPunished,
-    SafelyAborted,
-}
-
-impl fmt::Display for BobState {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            BobState::Started { .. } => write!(f, "started"),
-            BobState::Negotiated(..) => write!(f, "negotiated"),
-            BobState::BtcLocked(..) => write!(f, "btc is locked"),
-            BobState::XmrLocked(..) => write!(f, "xmr is locked"),
-            BobState::EncSigSent(..) => write!(f, "encrypted signature is sent"),
-            BobState::BtcRedeemed(..) => write!(f, "btc is redeemed"),
-            BobState::CancelTimelockExpired(..) => write!(f, "cancel timelock is expired"),
-            BobState::BtcCancelled(..) => write!(f, "btc is cancelled"),
-            BobState::BtcRefunded(..) => write!(f, "btc is refunded"),
-            BobState::XmrRedeemed => write!(f, "xmr is redeemed"),
-            BobState::BtcPunished => write!(f, "btc is punished"),
-            BobState::SafelyAborted => write!(f, "safely aborted"),
-        }
-    }
-}
 
 // TODO(Franck): Make this a method on a struct
 #[allow(clippy::too_many_arguments)]
