+++ swap/tests/happy_path_restart_alice.rs
@@ -7,7 +7,7 @@ use swap::{
     config::Config,
     database::Database,
     monero,
-    protocol::{alice, alice::swap::AliceState, bob},
+    protocol::{alice, alice::AliceState, bob},
 };
 use tempfile::tempdir;
 use testcontainers::clients::Cli;
