+++ swap/tests/testutils/mod.rs
@@ -9,7 +9,7 @@ use swap::{
     database::Database,
     monero,
     network::transport::build,
-    protocol::{alice, alice::swap::AliceState, bob, bob::swap::BobState},
+    protocol::{alice, alice::AliceState, bob, bob::BobState},
     SwapAmounts,
 };
 use tempfile::tempdir;
